backend/pyproject.toml
[tool.poetry]
name = "your-virtual-env"
version = "0.1.0"
description = "This repository contains the backend implementation of data loaders for the Capital Markets demo."
authors = ["Julian Boronat <julian.boronat@mongodb.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
pymongo = "^4.10.1"
python-dotenv = "^1.0.1"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
boto3 = "^1.35.70"
botocore = "^1.35.70"
scheduler = "^0.8.8"
tqdm = "^4.67.1"
transformers = "^4.49.0"
torch = "^2.6.0"
yfinance = "^0.2.54"
pyfredapi = "^0.9.1"
pandas = "^2.2.3"
voyageai = "^0.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
